added signed int

+++ 0x14-file_io/1-create_file.c
@@ -12,7 +12,7 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	size_t file;
+	signed int file;
 	int count = 0;
 
 	if (!filename)
